requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.9.24
charset-normalizer==2.1.1
Django==3.2.16
djangorestframework==3.14.0
idna==3.4
psycopg2==2.9.4
pycodestyle==2.9.1
pytz==2022.5
request2==0.2
requests==2.28.1
sqlparse==0.4.3
toml==0.10.2
urllib3==1.26.12
